inference_docker/requirements.txt
tqdm==4.62.3
scikit-image==0.19.2
openslide-python==1.1.2
torch==1.11.0
h5py==3.6.0
segmentation_models_pytorch==0.2.1
torchvision==0.12.0
albumentations==1.1.0
pandas==1.4.2
opencv-python==4.5.5.64
Shapely==1.8.2
GPUtil==1.4.0